+++ packages/in_app_purchase/lib/src/billing_client_wrappers/sku_details_wrapper.dart
@@ -42,7 +42,6 @@ class SkuDetailsWrapper {
     required this.subscriptionPeriod,
     required this.title,
     required this.type,
-    required this.isRewarded,
     required this.originalPrice,
     required this.originalPriceAmountMicros,
   });
@@ -106,10 +105,6 @@ class SkuDetailsWrapper {
   /// The [SkuType] of the product.
   final SkuType type;
 
-  /// False if the product is paid.
-  @JsonKey(defaultValue: false)
-  final bool isRewarded;
-
   /// The original price that the user purchased this product for.
   @JsonKey(defaultValue: '')
   final String originalPrice;
@@ -138,7 +133,6 @@ class SkuDetailsWrapper {
         typedOther.subscriptionPeriod == subscriptionPeriod &&
         typedOther.title == title &&
         typedOther.type == type &&
-        typedOther.isRewarded == isRewarded &&
         typedOther.originalPrice == originalPrice &&
         typedOther.originalPriceAmountMicros == originalPriceAmountMicros;
   }
@@ -158,7 +152,6 @@ class SkuDetailsWrapper {
         subscriptionPeriod.hashCode,
         title.hashCode,
         type.hashCode,
-        isRewarded.hashCode,
         originalPrice,
         originalPriceAmountMicros);
   }