+++ packages/in_app_purchase/lib/src/in_app_purchase/google_play_connection.dart
@@ -81,8 +81,7 @@ class GooglePlayConnection
   }
 
   @override
-  Future<BillingResultWrapper> completePurchase(PurchaseDetails purchase,
-      {String? developerPayload}) async {
+  Future<BillingResultWrapper> completePurchase(PurchaseDetails purchase) async {
     if (purchase.billingClientPurchase!.isAcknowledged) {
       return BillingResultWrapper(responseCode: BillingResponse.ok);
     }
@@ -91,20 +90,17 @@ class GooglePlayConnection
           'completePurchase unsuccessful. The `purchase.verificationData` is not valid');
     }
     return await billingClient.acknowledgePurchase(
-        purchase.verificationData.serverVerificationData,
-        developerPayload: developerPayload);
+        purchase.verificationData.serverVerificationData);
   }
 
   @override
-  Future<BillingResultWrapper> consumePurchase(PurchaseDetails purchase,
-      {String? developerPayload}) {
+  Future<BillingResultWrapper> consumePurchase(PurchaseDetails purchase) {
     if (purchase.verificationData == null) {
       throw ArgumentError(
           'consumePurchase unsuccessful. The `purchase.verificationData` is not valid');
     }
     return billingClient.consumeAsync(
-        purchase.verificationData.serverVerificationData,
-        developerPayload: developerPayload);
+        purchase.verificationData.serverVerificationData);
   }
 
   @override