+++ packages/in_app_purchase/test/in_app_purchase_connection/google_play_connection_test.dart
@@ -633,8 +633,7 @@ void main() {
       if (purchaseDetails.pendingCompletePurchase) {
         final BillingResultWrapper billingResultWrapper =
             await GooglePlayConnection.instance.completePurchase(
-                purchaseDetails,
-                developerPayload: 'dummy payload');
+                purchaseDetails);
         expect(billingResultWrapper, equals(expectedBillingResult));
         completer.complete(billingResultWrapper);
       }